requirements.txt
# Core LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0

# Ollama integration
langchain-ollama>=0.1.0

# Tavily search tool
tavily-python>=0.3.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=5.0.0

# Environment variables
python-dotenv>=1.0.0

# Data handling / utilities
pydantic>=2.0.0
tqdm>=4.66.0

# Optional: async + performance
aiohttp>=3.9.0
httpx>=0.27.0

# Optional: vector store (if you add memory retrieval later)
chromadb>=0.5.0

# Optional: embeddings (if needed later)
sentence-transformers>=2.6.0

# Logging / debugging
loguru>=0.7.0

streamlit